pyBackend/requirements.txt
fastapi
uvicorn
pillow
opencv-python
easyocr
fuzzywuzzy
python-Levenshtein
